Refactor NotesTree component for improved structure and functionality
- Removed unused imports and consolidated state management for clarity.
- Introduced new components for better organization: NotesTreeHeader, TreeNodeComponents, and useNotesTreeStructure.
- Enhanced drag-and-drop functionality with useNotesTreeDragDrop hook.
- Implemented type definitions in NotesTree.types.ts for better type safety.
- Improved search and filter capabilities in the NotesTree, allowing for a more intuitive user experience.
- Cleaned up the layout and styling for better responsiveness and usability.

This refactor aims to enhance maintainability and performance while providing a more organized structure for the NotesTree component.
